requirements.txt
langchain-gigachat
loguru~=0.6.0
langchain-core~=0.3.25
python-dotenv~=1.0.1